Revert "Db relationship"
