+++ server/seed/Language.json
@@ -1,0 +1,22 @@
+{
+	"languages": [
+		{ "name": "Java" },
+		{ "name": "JavaScript" },
+		{ "name": "Python" },
+		{ "name": "C++" },
+		{ "name": "Golang" },
+		{ "name": "C#" },
+		{ "name": "PHP" },
+		{ "name": "Ruby" },
+		{ "name": "Objective-C" },
+		{ "name": "SQL" },
+		{ "name": "Swift" },
+		{ "name": "TypeScript" },
+		{ "name": "HTML" },
+		{ "name": "CSS" },
+		{ "name": "Dart" },
+		{ "name": "React" },
+		{ "name": "Angular" },
+		{ "name": "Kotlin" }
+	]
+}
